Calendar Additions
made a hovering menu when you hover over schedule you can choose either students or teachers. If you click on one of them it'll show you a dummy list of teachers and students. There is a working search bar where you can search the students that you want to find. If you click on any of the students it brings you to a page where a blank calendar is. As an admin you are able to add and delete things on the calendar. Also added a sample teacher user who can also see their own calendar.